feat: refactoring create_environment_values_file to module structure
