+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 
 
-from .script_create_vars_file import ScriptCreateVarsFile
+from .script_create_environment_values_file import ScriptEnvironValuesFileCreator
+from .script_create_vars_file import ScriptVarsFileCreator
 
 
 __all__ = [
-    "ScriptCreateVarsFile",
+    "ScriptEnvironValuesFileCreator",
+    "ScriptVarsFileCreator",
 ]
