+++ app/v2/main/helm_chart_helper_v1.py
@@ -474,9 +474,9 @@ def create_vars_file_version1(conf):
     create_vars_file_version1
     """
 
-    from app.v1.modules.kubernetes_manager import ScriptCreateVarsFile
+    from app.v1.modules.kubernetes_manager import ScriptVarsFileCreator
 
-    ScriptCreateVarsFile.create_vars_file(conf)
+    ScriptVarsFileCreator.create_vars_file(conf)
 
 
 def create_vars_file(config_settings):
@@ -572,54 +572,28 @@ def create_helm_chart(config_settings):
     pass
 
 
-def create_environment_values_file(conf):
-    file = f"config_files/output/environments/{conf['chart']['name']}/{conf['chart']['name']}.values.yaml.gotmpl"
-    data = list()
-    data.append('# Load global and environment settings')
-    data.append('{{ readFile "../../include/values.global.yaml" }}')
-    data.append('{{ readFile "values.environment.yaml" }}')
-    data.append('')
-    for kind in conf['kubernetes']['values'].keys():
-        if kind == 'ConfigMap':
-            data.append(f'### {kind}')
-            for component in conf['kubernetes']['values'][kind]:
-                data.append(f'{component}:')
-                data.append(f'  data: {{}}')
-                # Save variables to file
-                data.append(f'')
-                pass
-        elif kind == 'Secret':
-            data.append(f'### {kind}')
-            for component in conf['kubernetes']['values'][kind]:
-                data.append(f'{component}:')
-                data.append(f'  stringData: {{}}')
-                data.append(f'')
-                pass
-        elif kind in ['Deployment', 'StatefulSet']:
-            data.append(f'### {kind}')
-            for component in conf['kubernetes']['values'][kind]:
-                data.append(f'{component}:')
-                data.append(f'  <<: *default-environment')
-                data.append(f'  <<: *default-global-resources-nolimit')
-                data.append(f"  replicaCount: {conf['kubernetes']['values'][kind][component]['replicas']}")
-                data.append(f"  image:")
-                data.append(f"    repository: {conf['kubernetes']['values'][kind][component]['image']['repository']}")
-                data.append(f"    tag: {conf['kubernetes']['values'][kind][component]['image']['tag']}")
-                data.append(f'')
-                pass
-        elif kind == 'Ingress':
-            data.append(f'### {kind}')
-            for component in conf['kubernetes']['values'][kind]:
-                data.append(f'{component}:')
-                if conf['kubernetes']['values'][kind][component]['annotations']:
-                    data.append(f'  annotations:')
-                    for key, value in conf['kubernetes']['values'][kind][component]['annotations'].items():
-                        data.append(f"    {key}: {value}")
-                data.append(f"  host: {conf['kubernetes']['values'][kind][component]['rules'][0]['host']}")
-                pass
-            data.append('')
-    write_file(file, data, mode='raw')
-    pass
+def create_environment_values_file_version1(conf):
+    """
+    create_environment_values_file_version1
+    """
+
+    from app.v1.modules.kubernetes_manager import ScriptEnvironValuesFileCreator
+
+    ScriptEnvironValuesFileCreator.create_environment_values_file(conf)
+
+
+def create_environment_values_file(config_settings):
+    """
+    create_environment_values_file
+    """
+
+    global app_version
+
+    if app_version == "version1":
+        create_environment_values_file_version1(config_settings)
+
+    if app_version == "version21":
+        create_environment_values_file_version1(config_settings)
 
 
 class AppMainManager:
